reduce amount of setup_*.py files in toplevel dir
